requirements.txt
Flask==3.0.3
gunicorn==21.2.0
google-cloud-vision==3.7.4
google-cloud-aiplatform==1.66.0
google-auth[requests]==2.34.0
google-cloud-firestore>=2.15.0
numpy==1.26.4
Pillow==10.4.0
opencv-python-headless==4.10.0.84
onnxruntime==1.18.1